Rename archived/featured listings to RESTful index controllers
Both extracted listings used an adjective-named controller with a non-RESTful
show action. Turn them into plural resources exposing index:

- Admin::JobOffers::ArchivedController#show -> ArchivesController#index
  (resources :archives, only: :index).
- Merge the featured listing into the existing FeaturesController as its
  index action, next to create/destroy; drop the now-redundant
  FeaturedController and move its view to features/index.html.haml.

Update the routes, navbar links, the archived_listing?/featured_listing?
helpers, and rename/merge the request specs accordingly.
