+++ app/helpers/admin/job_offers_helper.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 module Admin::JobOffersHelper
-  def archived_listing? = controller.is_a?(Admin::JobOffers::ArchivedController)
+  def archived_listing? = controller.is_a?(Admin::JobOffers::ArchivesController)
 
-  def featured_listing? = controller.is_a?(Admin::JobOffers::FeaturedController)
+  def featured_listing? = controller.is_a?(Admin::JobOffers::FeaturesController)
 
   def charts_per_day_options
     {
