+++ app/views/admin/job_offers/_navbar.html.haml
@@ -12,11 +12,11 @@
   .list-group.stepper
     - klasses = %w(list-group-item d-flex justify-content-between align-items-center)
     - klasses << 'active' if archived_state_activated
-    = link_to admin_job_offers_archived_path, class: klasses do
+    = link_to admin_job_offers_archives_path, class: klasses do
       Offres archivées
   - if can? :feature, :job_offer
     .list-group.stepper
       - klasses = %w(list-group-item d-flex justify-content-between align-items-center)
       - klasses << 'active' if featured_state_activated
-      = link_to admin_job_offers_featured_path, class: klasses do
+      = link_to admin_job_offers_features_path, class: klasses do
         Offres à la une
