+++ config/routes.rb
@@ -71,8 +71,8 @@ Rails.application.routes.draw do
         post :init, to: "job_offers#new"
         get :init, to: "job_offer_terms#index"
         get :add_actor
-        resource :featured, only: :show, controller: "job_offers/featured", as: :job_offers_featured
-        resource :archived, only: :show, controller: "job_offers/archived", as: :job_offers_archived
+        resources :features, only: :index, controller: "job_offers/features", as: :job_offers_features
+        resources :archives, only: :index, controller: "job_offers/archives", as: :job_offers_archives
         JobOffer.aasm.events.map(&:name).each do |event_name|
           action_name = :"create_and_#{event_name}"
           post :create, constraints: CommitParamConstraint.new(action_name), action: action_name
