+++ spec/requests/admin/job_offers/archives_spec.rb
@@ -2,9 +2,9 @@
 
 require "rails_helper"
 
-RSpec.describe "Admin::JobOffers::Archived" do
-  describe "GET /admin/offresdemploi/archived" do
-    subject(:archived_request) { get admin_job_offers_archived_path }
+RSpec.describe "Admin::JobOffers::Archives" do
+  describe "GET /admin/offresdemploi/archives" do
+    subject(:archives_request) { get admin_job_offers_archives_path }
 
     before do
       sign_in create(:administrator)
@@ -12,7 +12,7 @@ RSpec.describe "Admin::JobOffers::Archived" do
     end
 
     context "when the administrator can read job offers" do
-      before { archived_request }
+      before { archives_request }
 
       it { expect(response).to be_successful }
     end
@@ -20,7 +20,7 @@ RSpec.describe "Admin::JobOffers::Archived" do
     context "when the administrator cannot read job offers" do
       before do
         allow_any_instance_of(Ability).to receive(:can?).and_return(false)
-        archived_request
+        archives_request
       end
 
       it { expect(response).to have_http_status(:forbidden) }
